SnapAid-model/requirements.txt
transformers
torch
pandas
scikit-learn
google-generativeai
fastapi
uvicorn
dotenv